refactor(bindings): update nrs bindings, remove extra function from keys
